HLE: move SharedPage into Kernel
similar to config_mem, kernel is responsible for setting up this
