+++ src/core/core.h
@@ -8,7 +8,6 @@
 #include <string>
 #include "common/common_types.h"
 #include "core/frontend/applets/swkbd.h"
-#include "core/hle/shared_page.h"
 #include "core/loader/loader.h"
 #include "core/memory.h"
 #include "core/perf_stats.h"
@@ -203,10 +202,6 @@ public:
         return registered_swkbd;
     }
 
-    std::shared_ptr<SharedPage::Handler> GetSharedPageHandler() const {
-        return shared_page_handler;
-    }
-
 private:
     /**
      * Initialize the emulated system.
@@ -245,9 +240,6 @@ private:
     /// RPC Server for scripting support
     std::unique_ptr<RPC::RPCServer> rpc_server;
 #endif
-
-    /// Shared Page
-    std::shared_ptr<SharedPage::Handler> shared_page_handler;
 
     std::unique_ptr<Service::FS::ArchiveManager> archive_manager;
