+++ src/core/hle/kernel/shared_page.cpp
@@ -5,8 +5,8 @@
 #include <chrono>
 #include <cstring>
 #include "core/core_timing.h"
+#include "core/hle/kernel/shared_page.h"
 #include "core/hle/service/ptm/ptm.h"
-#include "core/hle/shared_page.h"
 #include "core/movie.h"
 #include "core/settings.h"
