+++ src/core/hle/kernel/shared_page.h
